fit. 게시글 조회 testcode 수정

+++ src/main/java/com/api/booklog/service/PostService.java
@@ -100,7 +100,11 @@ public class PostService {
     public void delete(Long userId, Long postId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow((PostNotFound::new));
-        bookMarkService.removeBookmark(userId, postId);
+
+        boolean isBookmarked = bookMarkService.isMemberOfZSet(bookMarkService.makeKey(userId), postId);
+        if(isBookmarked) {
+            bookMarkService.removeBookmark(userId, postId);
+        }
         postRepository.delete(post);
     }
